// src/types/api/service/models/CreateClipBookDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateClipBookDto
 */
export interface CreateClipBookDto {
    /**
     * 이름
     * @type {string}
     * @memberof CreateClipBookDto
     */
    title: string;
    /**
     * 클립북 유형, 공통코드 - CLIPBOOK_TYPE : NORMAL(일반 클립북) / COVERAGE(커버리지 클립북)
     * @type {string}
     * @memberof CreateClipBookDto
     */
    type?: string;
    /**
     * 프로젝트ID 목록, N차 개발 예정
     * @type {Array<number>}
     * @memberof CreateClipBookDto
     */
    projectIdList?: Array<number>;
    /**
     * 현재 선택된 사용자의 그룹ID
     * @type {number}
     * @memberof CreateClipBookDto
     */
    groupId: number;
    /**
     * 공유설정
     * @type {string}
     * @memberof CreateClipBookDto
     */
    shareCode?: string;
    /**
     * 언론인목록 공유설정 기본값설정여부
     * @type {boolean}
     * @memberof CreateClipBookDto
     */
    chkDefault?: boolean;
    /**
     * 뉴스 목록
     * @type {Array<number>}
     * @memberof CreateClipBookDto
     */
    newslist?: Array<number>;
    /**
     * 뉴스 인덱스 목록(YYYYMM)
     * @type {Array<string>}
     * @memberof CreateClipBookDto
     */
    newsIndexlist?: Array<string>;
    /**
     * 보도자료 목록
     * @type {Array<number>}
     * @memberof CreateClipBookDto
     */
    prlist?: Array<number>;
}

/**
 * Check if a given object implements the CreateClipBookDto interface.
 */
export function instanceOfCreateClipBookDto(value: object): value is CreateClipBookDto {
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('groupId' in value) || value['groupId'] === undefined) return false;
    return true;
}

export function CreateClipBookDtoFromJSON(json: any): CreateClipBookDto {
    return CreateClipBookDtoFromJSONTyped(json, false);
}

export function CreateClipBookDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateClipBookDto {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'],
        'type': json['type'] == null ? undefined : json['type'],
        'projectIdList': json['projectIdList'] == null ? undefined : json['projectIdList'],
        'groupId': json['groupId'],
        'shareCode': json['shareCode'] == null ? undefined : json['shareCode'],
        'chkDefault': json['chkDefault'] == null ? undefined : json['chkDefault'],
        'newslist': json['newslist'] == null ? undefined : json['newslist'],
        'newsIndexlist': json['newsIndexlist'] == null ? undefined : json['newsIndexlist'],
        'prlist': json['prlist'] == null ? undefined : json['prlist'],
    };
}

export function CreateClipBookDtoToJSON(value?: CreateClipBookDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'title': value['title'],
        'type': value['type'],
        'projectIdList': value['projectIdList'],
        'groupId': value['groupId'],
        'shareCode': value['shareCode'],
        'chkDefault': value['chkDefault'],
        'newslist': value['newslist'],
        'newsIndexlist': value['newsIndexlist'],
        'prlist': value['prlist'],
    };
}